refactor and edit account routing
