+++ src/AwesomeApp.UI/src/common/types/NextAuth.ts
@@ -8,7 +8,7 @@ declare module 'next-auth' {
   }
 
   interface Session extends DefaultSession {
-    user?: User;
+    user: User;
   }
 }
