+++ src/AwesomeApp.UI/src/frontend/components/index.ts
@@ -1,1 +1,3 @@
 export * from './page-box'
+export * from './snackbar'
+export * from './spinner'
