+++ src/AwesomeApp.UI/src/frontend/components/spinner/Spinner.tsx
