+++ src/AwesomeApp.UI/src/frontend/components/spinner/styles.module.css
