--- src/AwesomeApp.UI/src/pages/_components/index.ts
@@ -1,2 +1,0 @@
-export { SnackbarProvider } from './snackbar'
-export { SpinnerProvider } from './spinner'
