+++ src/AwesomeApp.UI/src/pages/_hooks/index.ts
@@ -1,3 +1,3 @@
 export { useCallWithErrorHandling, useFetchWithErrorHandling } from './ApiCallWithErrorHandling'
-export { useSpinner } from '@/pages/_components/spinner'
-export { useSnackbar } from '@/pages/_components/snackbar'
+export { useSpinner } from '../../frontend/components/spinner'
+export { useSnackbar } from '../../frontend/components/snackbar'
