+++ src/AwesomeApp.UI/src/pages/account/_editAccount/index.tsx
@@ -1,0 +1,11 @@
+import { Account } from "@/common/types/account"
+
+type Props = {
+  account: Account
+}
+
+export default function EditAccountComponent({ account }: Readonly<Props>) {
+  return (
+    <>{account.id}</>
+  )
+}
