+++ src/AwesomeApp.UI/src/pages/account/edit/index.ts
@@ -1,0 +1,2 @@
+export { getServerSideProps } from './_serverProps'
+export { default } from '../_editAccount'
